description_gettingmoney.py
write a program that prints a paycheck. Ask the program user for the name of the employee, the hourly rate, and the number of hours worked. If the number of hours exceeds 40, the employee is paid "time and a half", that is, 150 percent of the hourly rate on hours exceeding 40. Use fictitious names for the payer and the bank.